Removing some warning and making things more clear in general. AKA less unicode2char

+++ src/AICHSyncThread.cpp
@@ -222,7 +222,7 @@ void* CAICHSyncThread::Entry()
 	//	}
 
 	if ( !queue.empty() ) {
-		printf("AICH Thread: Starting to hash files. %d files found.\n", queue.size() );
+		printf("AICH Thread: Starting to hash files. %li files found.\n", (long int)queue.size() );
 		while ( !queue.empty() ) {
 			// Check for termination
 			if ( TestDestroy() ) {
@@ -232,8 +232,8 @@ void* CAICHSyncThread::Entry()
 			CKnownFile* pCurFile = queue.front();
 			queue.pop_front();
 
-			printf("AICH Thread: Hashing file: %s, total files left: %d\n",
-				(const char *)unicode2char( pCurFile->GetFileName() ), queue.size() );
+			printf("AICH Thread: Hashing file: %s, total files left: %li\n",
+				(const char *)unicode2char( pCurFile->GetFileName() ), (long int)queue.size() );
 
 			// Just to be sure that the file hasnt been deleted lately
 			if ( 	!(theApp.knownfiles->IsKnownFile(pCurFile) &&
